requirements.txt
requests
beautifulsoup4
unidecode
spotipy
python-dotenv
